+++ src/core/teon/index.rs
@@ -97,7 +97,7 @@ struct Type<'a>(&'a Value);
 
 impl<'a> Display for Type<'a> {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        match *self.0 {
+        match self.0 {
             Value::Null => formatter.write_str("null"),
             Value::Bool(_) => formatter.write_str("bool"),
             #[cfg(feature = "data-source-mongodb")]
@@ -118,7 +118,7 @@ impl<'a> Display for Type<'a> {
             Value::Range(_) => formatter.write_str("range"),
             Value::Tuple(_) => formatter.write_str("tuple"),
             Value::Pipeline(_) => formatter.write_str("pipeline"),
-            Value::RawEnumChoice(_) => formatter.write_str("raw enum choice"),
+            Value::RawEnumChoice(_, __) => formatter.write_str("raw enum choice"),
             Value::RawOptionChoice(_) => formatter.write_str("raw option choice"),
             Value::RegExp(_) => formatter.write_str("regular expression"),
         }